# specs/001-grooming-saas-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Grooming SaaS API
  version: 0.1.0
  description: |
    Contrato REST del MVP de SaaS Peluquería Canina.
    Fuente de verdad compartida entre `api-micro-saas` (backend NestJS) y
    `www-micro-saas` (frontend Next.js). Toda escritura está scoped por el
    `tenantId` extraído del JWT — el cliente NUNCA envía `tenantId` en el body.
servers:
  - url: http://localhost:3001
    description: Local dev (backend NestJS por defecto)
  - url: https://api.grooming.example.com
    description: Producción
...
security:
  - bearerAuth: []
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: Listar citas (filtro por fecha opcional)
      parameters:
        - in: query
          name: date
          schema:
            type: string
            format: date
          description: Si se omite, devuelve las citas del día actual.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
    post:
      tags: [Appointments]
      summary: Crear cita
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateInput'
      responses:
        '201':
          description: Cita creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Payload inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    Appointment:
      type: object
      required:
        - id
        - clientId
        - petId
        - scheduledAt
        - service
        - reminderStatus
      properties:
        id: { type: string, format: uuid }
        clientId: { type: string, format: uuid }
        petId: { type: string, format: uuid }
        clientName: { type: string }
        petName: { type: string }
        scheduledAt: { type: string, format: date-time }
        service: { type: string }
        reminderStatus:
          type: string
          enum: [not_sent, sent]
        reminderSentAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
    AppointmentCreateInput:
      type: object
      required: [clientId, petId, scheduledAt, service]
      properties:
        clientId: { type: string, format: uuid }
        petId: { type: string, format: uuid }
        scheduledAt: { type: string, format: date-time }
        service: { type: string, minLength: 1 }
...
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string }